feat: update directory
